requirements.txt
# Core Dependencies
python-dotenv==1.0.0

# Database
supabase>=2.5.0

# Data Processing & Analysis
pandas
numpy

# Machine Learning (compatible with NumPy 2.x)
scikit-learn>=1.4

# Visualization
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.17.0

# Frontend
streamlit==1.28.1

# API & HTTP
requests==2.31.0
httpx>=0.26

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Explainability
shap==0.45.1

# LLM / AI Integration
groq>=0.4.1

# File Processing
pdfplumber>=0.10.3
openpyxl>=3.1.2
xlrd>=2.0.1

# Config / YAML
pyyaml>=6.0.1